Remove unused gcs and dc endpoints. (#170)
The only known caller of the coin supply and download count endpoints was JavaScript running on decred.org. Since the site overhaul, these endpoints are not used.

Further justifying their removal is the fact that coin supply is readily available from dcrdata API, and the download count was a very rough estimate based on a quite basic heuristic.